Greatly improve service structure
