+++ scripts/utils/button_service.py
@@ -5,21 +5,12 @@ import uasyncio
 
 
 class ButtonService:
-    def __init__(
-        self,
-        view_manager,
-        views,
-        current_view_task,
-        galactic_unicorn,
-        graphics,
-        sound_service,
-    ):
+    def __init__(self, view_service, current_view_task):
         self.current_view_task = current_view_task
-        self.galactic_unicorn = galactic_unicorn
-        self.graphics = graphics
-        self.sound_service = sound_service
-        self.view_manager = view_manager
-        self.views = views
+        self.galactic_unicorn = view_service.galactic_unicorn
+        self.sound_service = view_service.sound_service
+        self.view_service = view_service
+        self.views = view_service.get_views()
 
         # Initialize button states
         self.button_states = {
@@ -64,16 +55,13 @@ class ButtonService:
             if not self.button_states["A"]:
                 self.button_states["A"] = True
                 await self.sound_service.stop_all_sounds()
-                current_index = view_keys.index(self.view_manager.current_view_key)
-                self.view_manager.current_view_key = view_keys[
+                current_index = view_keys.index(self.view_service.current_view_key)
+                self.view_service.current_view_key = view_keys[
                     (current_index - 1) % len(view_keys)
                 ]
-                self.current_view_task = await self.view_manager.switch_view(
+                self.current_view_task = await self.view_service.switch_view(
                     self.views,
                     self.current_view_task,
-                    self.galactic_unicorn,
-                    self.graphics,
-                    self.sound_service,
                 )
         else:
             self.button_states["A"] = False
@@ -83,16 +71,13 @@ class ButtonService:
             if not self.button_states["B"]:
                 self.button_states["B"] = True
                 await self.sound_service.stop_all_sounds()
-                current_index = view_keys.index(self.view_manager.current_view_key)
-                self.view_manager.current_view_key = view_keys[
+                current_index = view_keys.index(self.view_service.current_view_key)
+                self.view_service.current_view_key = view_keys[
                     (current_index + 1) % len(view_keys)
                 ]
-                self.current_view_task = await self.view_manager.switch_view(
+                self.current_view_task = await self.view_service.switch_view(
                     self.views,
                     self.current_view_task,
-                    self.galactic_unicorn,
-                    self.graphics,
-                    self.sound_service,
                 )
         else:
             self.button_states["B"] = False
@@ -160,9 +145,7 @@ class ButtonService:
                     self.previous_brightness = self.galactic_unicorn.get_brightness()
                     self.galactic_unicorn.set_brightness(0)
                     await self.sound_service.toggle_mute()
-                    self.graphics.set_pen(self.graphics.create_pen(0, 0, 0))
-                    self.graphics.clear()
-                    self.galactic_unicorn.update(self.graphics)
+                    self.view_service.clear_screen()
                     self.is_device_on = False
                 else:
                     self.galactic_unicorn.set_brightness(self.previous_brightness)
